Refactor and clean up backend implementations

+++ kaze/platform/backends/sdl3/GamepadConstants.inl
@@ -1,3 +1,6 @@
+/// @file GamepadConstants.inl
+/// @description
+/// Contains gamepad constants, which map SDL constants with kaze constants
 #pragma once
 #ifndef kaze_platform_backends_sdl3_gamepadconstants_inl_
 #define kaze_platform_backends_sdl3_gamepadconstants_inl_
@@ -8,7 +11,7 @@
 
 KAZE_NAMESPACE_BEGIN
 
-static Array<Uint8, static_cast<Size>(GamepadBtn::Count)> s_gamepadButtonToSDL = {
+static Array<int, static_cast<Size>(GamepadBtn::Count)> s_gamepadButtonToSDL = {
     SDL_GAMEPAD_BUTTON_SOUTH,
     SDL_GAMEPAD_BUTTON_EAST,
     SDL_GAMEPAD_BUTTON_WEST,
@@ -26,7 +29,7 @@ static Array<Uint8, static_cast<Size>(GamepadBtn::Count)> s_gamepadButtonToSDL =
     SDL_GAMEPAD_BUTTON_DPAD_DOWN,
 };
 
-static Array<Uint8, static_cast<Size>(GamepadAxis::Count)> s_gamepadAxisToSDL = {
+static Array<int, static_cast<Size>(GamepadAxis::Count)> s_gamepadAxisToSDL = {
     SDL_GAMEPAD_AXIS_LEFTX,
     SDL_GAMEPAD_AXIS_LEFTY,
     SDL_GAMEPAD_AXIS_RIGHTX,
@@ -35,8 +38,8 @@ static Array<Uint8, static_cast<Size>(GamepadAxis::Count)> s_gamepadAxisToSDL = 
     SDL_GAMEPAD_AXIS_RIGHT_TRIGGER,
 };
 
-static Array<Uint8, SDL_GAMEPAD_BUTTON_COUNT> s_sdlToGamepadButton = {};
-static Array<Uint8, SDL_GAMEPAD_AXIS_COUNT> s_sdlToGamepadAxis = {};
+static Array<int, SDL_GAMEPAD_BUTTON_COUNT> s_sdlToGamepadButton = {};
+static Array<int, SDL_GAMEPAD_AXIS_COUNT> s_sdlToGamepadAxis = {};
 
 KAZE_NAMESPACE_END
